feat: Add the function for extracting the frequency features (#13)
* feat: Add a function to calculate MNF, MDF, VCF

Add a function to calculate MNF, MDF, VCF.
But, there's no exception, docstring.

* feat: Add exceptions for '_get_amp_and_freq' func

The exceptions are for type and value of parameters.

Meanwhile, the logic is modified to calculate the valid 'low_idx' and
'high idx' of freq_range.

e.g. If the 'freq_range' is (10, 0), it would have bugs.

* feat: Add function for multi-dimensional x

In this commit, I considered caculations for multi-dimensional x.
So, we can calculate mnf, mdf, vcf for multi-dimensinal x.
But, the code was not tested. So, there may be a bug in the code.

* fix: Fix the bugs for mnf, mdf

For mnf, there were the bugs; exception not working and
when axis is -1, unwanted result comes out.
For mdf, The function calculated the amplitude of the median
frequency without calculating the median frequency.

* fix: Modify the bug of the mdf. Add keepdims

Make the median frequency calculation possible along the `axis`.
And, Adds the function of maintaining the dimension of the function
calculation result.

* test: Add test code for frequency_feature.py

Tests for the mnf, mdf, vcf function when the input is the 1-D array or
2-D arrays.

* fix: Modify the __init__.py file

Modify the file to merge to main branch.

* refactor: Change the code structure for mdf

Change the code structure for readability and the function name.

* docs: Add docstrings for `mnf`, `mdf`, `vcf`

Docstrings includes the information; summary, parameters, returns,
examples etc.

* feat: Change an input parameter of function

The input parameter `x` of `mnf`, `mdf`, `vcf` changed to `amp`.
This means the function needs the amplitudes of spectrum not the time
-domain signals. The reason for the change is as follows;
when you use the function several times, the time signal should be done
FFT for each function execution.

* feat: Change the structure of the folder

- Change the name from frequency_feature.py to frequency.py
- Move the frequency.py from oplib.signal to oplib.feature.
- Add doctrings for the formula of `mnf`, `mdf`, `vcf`.

* docs: Changes to documentation for docstrings

Changes for the short summary